+++ components/03-modules/catalog-list-item/catalog-list-item.config.js
@@ -40,7 +40,7 @@ module.exports = {
       class: 'button--icon catalog-list-item__actions-secondary-button',
       iconId: 'heart',
       iconClass: 'button__icon catalog-list-item__actions-secondary-button-icon',
-      attributes: 'type="button" aria-label="button wishlist"'
+      attributes: 'type="button" aria-label="Add to wishlist"'
     },
     buttonCompare: {
       tag: 'button',
@@ -48,7 +48,7 @@ module.exports = {
       class: 'button--icon catalog-list-item__actions-secondary-button',
       iconId: 'compare',
       iconClass: 'button__icon catalog-list-item__actions-secondary-button-icon',
-      attributes: 'type="button" aria-label="button compare"'
+      attributes: 'type="button" aria-label="Add to compare"'
     }
   },
   variants: [