+++ components/03-modules/filter/filter.config.js
@@ -53,28 +53,28 @@ module.exports = {
         options: [
           {
             text: 'XS',
-            ariaLabel: 'Placaholder for arialabel',
+            ariaLabel: 'Select size XS',
             href: '#'
           },
           {
             class: 'filter__swatch-option--active',
             text: 'S',
-            ariaLabel: 'Placaholder for arialabel',
+            ariaLabel: 'Select size s',
             href: '#'
           },
           {
             text: 'M',
-            ariaLabel: 'Placaholder for arialabel',
+            ariaLabel: 'Select size M',
             href: '#'
           },
           {
             text: 'L',
-            ariaLabel: 'Placaholder for arialabel',
+            ariaLabel: 'Select size L',
             href: '#'
           },
           {
             text: 'XL',
-            ariaLabel: 'Placaholder for arialabel',
+            ariaLabel: 'Select size XL',
             href: '#'
           }
         ]
@@ -94,28 +94,28 @@ module.exports = {
             class: 'filter__color-option--active',
             attributes: 'style="background: #000"',
             label: 'Black',
-            ariaLabel: 'Placeholder for aria-label'
+            ariaLabel: 'Select color black'
           },
           {
             href: '#',
             class: 'filter__color-option',
             attributes: 'style="background: red"',
             label: 'Red',
-            ariaLabel: 'Placeholder for aria-label'
+            ariaLabel: 'Select color red'
           },
           {
             href: '#',
             class: 'filter__color-option',
             attributes: 'style="background: #5cb85c"',
             label: 'Green',
-            ariaLabel: 'Placeholder for aria-label'
+            ariaLabel: 'Select color green'
           },
           {
             href: '#',
             class: 'filter__color-option',
             attributes: 'style="background: #0275d8"',
             label: 'Blue',
-            ariaLabel: 'Placeholder for aria-label'
+            ariaLabel: 'Select color blue'
           }
         ]
       }