+++ components/03-modules/minicart-content/minicart-content.config.js
@@ -12,6 +12,17 @@ module.exports = {
       class: 'button--secondary button--fluid minicart-content__checkout',
       attributes: 'type="button"',
       text: 'Go to Checkout'
-    }
+    },
+    products: [
+      {
+        attributes: {
+          script: false,
+        }
+      },
+      {
+        productName: 'Some product 2 name',
+        updateQty: false
+      }
+    ]
   }
 };