+++ components/03-modules/catalog-grid-item/catalog-grid-item.config.js
@@ -10,7 +10,6 @@ module.exports = {
       }
     },
     image: {
-      class: 'catalog-grid-item__image',
       dataSrc: '/images/catalog-grid-item/product-1_320_312.jpg'
     },
     name: 'Some product name - very long name because that\'s important',
@@ -39,18 +38,22 @@ module.exports = {
       iconId: 'shopping-cart',
       iconClass: 'icon catalog-grid-item__cart-icon'
     },
-    secondaryActions: [
-      {
-        icon: {
-          iconId: 'heart'
-        }
+    secondaryActions: {
+      wishlist: {
+        tag: 'button',
+        class: 'button--icon',
+        iconId: 'heart',
+        iconClass: 'button__icon',
+        attributes: 'type="button" aria-label="Add to Wish List"',
       },
-      {
-        icon: {
-          iconId: 'compare'
-        }
+      compare: {
+        tag: 'button',
+        class: 'button--icon',
+        iconId: 'compare',
+        iconClass: 'button__icon',
+        attributes: 'type="button" aria-label="Add to compare"',
       }
-    ],
+    },
     swatches: [
       {
         name: 'swatch',