+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Global focus style
 - Catalog grid item module
 - Catalog grid view
+- Modal module
 
 ### Changed
 - Footer component design