+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Catalog grid item module
 - Catalog grid view
 - Filters module.
+- Modal module
 
 ### Changed
 - Footer component design
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - List module styles
 - Newsletter module design
 - Refactored swatch element
+- Update label hidden style
 
 ### Removed
 - Old catalog product modules and catalog views