+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Catalog list view
 - Filters module.
 - Modal module
+- Working js for Quantity-update
 
 ### Changed
 - Rating - use svg icons as css background with svg-uri for encoding uri