+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Modal module
 - Working js for Quantity-update
 - Sticky footer
+- Minicart
 
 ### Changed
 - Rating - use svg icons as css background with svg-uri for encoding uri