+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Working js for Quantity-update
 - Sticky footer
 - Minicart
+- Forgot password view
 
 ### Changed
 - Rating - use svg icons as css background with svg-uri for encoding uri