+++ CHANGELOG.md
@@ -31,6 +31,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Login page, removed tabs and set two columns
 - Aria-label values
 - Registration view
+- Simplify filters
+- Remove unnecessary code from config
+- Update toolbar styles
 - Mega menu module to fit Snowdog Menu Module
 - Side menu module to fit Snowdog Menu Module