+++ CHANGELOG.md
@@ -31,8 +31,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Login page, removed tabs and set two columns
 - Aria-label values
 - Registration view
-- Simplify filters, remove unnecessary code from config
+- Simplify filters
+- Remove unnecessary code from config
 - Update toolbar styles
+- Mega menu module to fit Snowdog Menu Module
+- Side menu module to fit Snowdog Menu Module
 
 ### Removed
 - Old catalog product modules and catalog views